refactor: split github_access.rs into responsibility-based modules (#101)
Split the monolithic github_access.rs (729 lines) into 6 focused modules:
- types.rs: AccessResult, AccessError, GithubAccessChecker trait, DynGithubAccessChecker
- real.rs: RealGithubAccessChecker (production GitHub API calls)
- cached.rs: CachedGithubAccessChecker + cache constants + trait impl
- installation_sync.rs: fallback sync logic (impl CachedGithubAccessChecker)
- test_doubles.rs: 5 mock implementations for testing
- mod.rs: pub use re-exports maintaining full API compatibility

No behavior changes. All existing imports (crate::github_access::* and
boardflow_api::github_access::*) continue to work unchanged.
